This had to be ported from the master branch dude to RN Bob breaking everything. My apologies. Date of commit is off. Middle of the night woke up and went to fix.
Very early commit with the button view exposed to make it simple, due to RCT's 'demand' to work with typescript, need to port things over. Some issues with the ability to make the menu appear through method calls needs work, there's multiple exports that will be needed here of multiple classes, the view is easiest to make things just go. There are many things that need to be worked out with exporting multiple classes or just understanding better how RN's mandated usage of TSX works, can we can just ignore it? This has been delayed too long despite the fact we have working code right there, just needs to be bridged correctly. It could be possible to move some of these props in to the button itself but they should not be done that way. It's already taken a week to get RN's tooling and Xcode 12 to work correctly without a million extra issues, this weekend should hopefully find a resolution.	ad74441	Darvell Long <contact@darvell.me>	Oct 23, 2020 at 2:53 PM